+++ src/ExportAPI.jl
@@ -295,8 +295,8 @@ export
   setVariableRefence!,
   reshapeVec2Mat
 
-export ContinuousScalar
-export ContinuousEuclid
+export Position, Position1, Position2, Position3, Position4
+export ContinuousScalar, ContinuousEuclid # TODO figure out if this will be deprecated, Caesar.jl #807
 export Circular, Circle
 
 # serializing distributions